TFKUBE-222: Cleanup e2e tests (#109)
* TFKUBE-222: Cleanup the end2end test

* TFKUBE-222: Modified github action

* Add defer to uninstall so the cleanup will run when tests failed.

* Removed e2e test from github action. This will be added in TFKUBE-230
