+++ test/e2etest/bamboo_test.go
@@ -1,273 +1,13 @@
 package e2etest
 
 import (
-	"context"
-	"encoding/base64"
 	"flag"
 	"fmt"
-	"strings"
-	"testing"
-	"time"
-
-	awsSdk "github.com/aws/aws-sdk-go/aws"
-	"github.com/aws/aws-sdk-go/service/eks"
-	"github.com/gruntwork-io/terratest/modules/aws"
-	"github.com/gruntwork-io/terratest/modules/environment"
-	"github.com/gruntwork-io/terratest/modules/k8s"
-	"github.com/gruntwork-io/terratest/modules/retry"
-	"github.com/gruntwork-io/terratest/modules/terraform"
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"testing"
 )
 
 var customConfigFilename = flag.String("config", "", "Name of test environment config file")
-
-func TestBambooModule(t *testing.T) {
-
-	environmentConfig := GenerateConfigForProductE2eTest(t, "bamboo", *customConfigFilename)
-	tfOptions := GenerateTerraformOptions(environmentConfig.TerraformConfig, t)
-	kubectlOptions := GenerateKubectlOptions(environmentConfig.KubectlConfig, tfOptions, environmentConfig.EnvironmentName)
-
-	if *customConfigFilename == "" {
-		creationErr := CreateDirIfNotExist("artifacts")
-		require.NoError(t, creationErr)
-		saveErr := Save("artifacts/"+defaultConfigFilename, environmentConfig)
-		require.NoError(t, saveErr)
-	}
-
-	// copy any changes to test directory
-	if *customConfigFilename != "" {
-		copyErr := CopyDir("../../", environmentConfig.TerraformConfig.TestFolder)
-		require.NoError(t, copyErr)
-	}
-
-	terraform.InitAndApply(t, tfOptions)
-
-	// Manually add default tags to ASG and EC2 after the initial provisioning due to the node group resource limitation(See: https://github.com/terraform-aws-modules/terraform-aws-eks/issues/1558)
-	tagAsgResources(t, environmentConfig)
-
-	vpcOutput := getVpcOutput(t, tfOptions)
-
-	assertVPC(t, environmentConfig.AwsRegion, vpcOutput, environmentConfig)
-	assertEKS(t, environmentConfig.AwsRegion, vpcOutput, environmentConfig)
-	assertShareHomePV(t, tfOptions, environmentConfig.EnvironmentName, environmentConfig.Product)
-	assertShareHomePVC(t, tfOptions, environmentConfig.EnvironmentName, environmentConfig.Product)
-	assertBambooPod(t, kubectlOptions, environmentConfig.Product)
-	assertIngressAccess(t, environmentConfig.Product, environmentConfig.EnvironmentName, fmt.Sprintf("%v", environmentConfig.TerraformConfig.Variables["domain"]))
-	assertRDS(t, tfOptions, kubectlOptions, environmentConfig.AwsRegion, environmentConfig.Product)
-	assertBambooSecrets(t, kubectlOptions, environmentConfig.Product)
-	assertBambooAgentPod(t, kubectlOptions)
-}
-
-func assertVPC(t *testing.T, awsRegion string, vpcOutput VpcOutput, environmentConfig EnvironmentConfig) {
-	vpc := aws.GetVpcById(t, vpcOutput.Id, awsRegion)
-	resourceTags := environmentConfig.TerraformConfig.Variables["resource_tags"].(map[string]string)
-
-	assert.Equal(t, fmt.Sprintf("atlas-%s-vpc", environmentConfig.EnvironmentName), vpc.Name)
-	assert.Equal(t, resourceTags["resource_owner"], vpc.Tags["resource_owner"])
-	assert.Len(t, vpc.Subnets, 4)
-}
-
-func assertEKS(t *testing.T, awsRegion string, vpcOutput VpcOutput, environmentConfig EnvironmentConfig) {
-	resourceTags := environmentConfig.TerraformConfig.Variables["resource_tags"].(map[string]string)
-	session := GenerateAwsSession(awsRegion)
-	eksClient := eks.New(session)
-	describeClusterInput := &eks.DescribeClusterInput{
-		Name: awsSdk.String(fmt.Sprintf("atlas-%s-cluster", environmentConfig.EnvironmentName)),
-	}
-
-	eksInfo, err := eksClient.DescribeCluster(describeClusterInput)
-	require.NoError(t, err)
-
-	assert.Equal(t, resourceTags["resource_owner"], *(eksInfo.Cluster.Tags["resource_owner"]))
-	assert.Equal(t, vpcOutput.Id, *((*eksInfo).Cluster.ResourcesVpcConfig.VpcId))
-}
-
-func assertShareHomePV(t *testing.T, tfOptions *terraform.Options, environmentName string, product string) {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-	k8sClient := K8sDriver(t, tfOptions, environmentName)
-	pvClient := k8sClient.CoreV1().PersistentVolumes()
-
-	_, err := pvClient.Get(ctx, fmt.Sprintf("atlassian-dc-%s-share-home-pv", product), v1.GetOptions{})
-	require.NoError(t, err)
-}
-
-func assertShareHomePVC(t *testing.T, tfOptions *terraform.Options, environmentName string, product string) {
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
-	k8sClient := K8sDriver(t, tfOptions, environmentName)
-	pvcClient := k8sClient.CoreV1().PersistentVolumeClaims(product)
-
-	_, err := pvcClient.Get(ctx, fmt.Sprintf("atlassian-dc-%s-share-home-pvc", product), v1.GetOptions{})
-	require.NoError(t, err)
-}
-
-func assertBambooPod(t *testing.T, kubectlOptions *k8s.KubectlOptions, product string) {
-	podName := fmt.Sprintf("%s-0", product)
-	pod := k8s.GetPod(t, kubectlOptions, podName)
-	k8s.WaitUntilPodAvailable(t, kubectlOptions, podName, 5, 30*time.Second)
-	shareHomeVolume := SafeExtractShareHomeVolume(pod.Spec.Volumes)
-
-	assert.Equal(t, true, pod.Status.ContainerStatuses[0].Ready)
-	assert.Equal(t, fmt.Sprintf("atlassian-dc-%s-share-home-pvc", product), shareHomeVolume.PersistentVolumeClaim.ClaimName)
-
-	// Make sure Bamboo is fully pre-seeded with setup wizard skipped. (<setupStep> will have 'complete' value)
-	output, execErr := k8s.RunKubectlAndGetOutputE(t, kubectlOptions, "exec", podName,
-		"--", "cat", "/var/atlassian/application-data/bamboo/bamboo.cfg.xml")
-	assert.Nil(t, execErr)
-
-	startIndex := strings.Index(output, "<setupStep>")
-	endIndex := strings.Index(output, "</setupStep>")
-	setupStep := output[startIndex+11 : endIndex]
-	assert.Equal(t, "complete", setupStep)
-}
-
-func assertIngressAccess(t *testing.T, product string, environment string, domain string) {
-	url := fmt.Sprintf("https://%s.%s.%s", product, environment, domain)
-	content := getPageContent(t, url)
-	expectedContent := "Time for an agent!"
-	assert.Contains(t, string(content), expectedContent)
-}
-
-func assertRDS(t *testing.T, tfOptions *terraform.Options, kubectlOptions *k8s.KubectlOptions, awsRegion string, product string) {
-	dbOutput := databaseOutput{}
-	terraform.OutputStruct(t, tfOptions, "database", &dbOutput)
-	dbInstanceID := dbOutput.RdsInstanceId
-	dbName := dbOutput.DbName
-
-	endpoint := aws.GetAddressOfRdsInstance(t, dbInstanceID, awsRegion)
-	port := aws.GetPortOfRdsInstance(t, dbInstanceID, awsRegion)
-
-	assert.NotNil(t, endpoint)
-	assert.Equal(t, int64(5432), port)
-
-	// Get password
-	secretName := product + "-db-cred"
-	secret, secretErr := k8s.RunKubectlAndGetOutputE(t, kubectlOptions, "get", "secret", secretName, "-o", "jsonpath='{.data.password}'")
-	assert.Nil(t, secretErr)
-	assert.NotNil(t, secret)
-	decSecret, decodeErr := base64.StdEncoding.DecodeString(secret[1 : len(secret)-1])
-	assert.Nil(t, decodeErr)
-	password := string(decSecret)
-
-	// Assert DB connection
-	psqlClientPodName := "e2e-test-psqlclient"
-	username := product + "user"
-
-	defer k8s.RunKubectlAndGetOutputE(t, kubectlOptions, "delete", "pod", psqlClientPodName)
-
-	_, psqlClientErr := k8s.RunKubectlAndGetOutputE(t, kubectlOptions, "run", psqlClientPodName,
-		"--image=tmaier/postgresql-client", "--command", "--", "/bin/sh", "-c", "tail -f /dev/null")
-	if psqlClientErr != nil && strings.Contains(psqlClientErr.Error(), "AlreadyExists") {
-		psqlClientErr = nil
-	}
-	assert.Nil(t, psqlClientErr)
-
-	ExpectedStatus := "success"
-	status := retry.DoWithRetry(t, "Waiting for DB connection validation...", 5, time.Duration(time.Second*5),
-		func() (string, error) {
-			output, execErr := k8s.RunKubectlAndGetOutputE(t, kubectlOptions, "exec", psqlClientPodName,
-				"--", "/bin/sh",
-				"-c", fmt.Sprintf("PGPASSWORD=\"%s\" psql \"sslmode=require host=%s dbname=%s user=%s\" -q -c \"SELECT version()\" > test.log;echo $?;",
-					password, endpoint, dbName, username),
-			)
-			if execErr != nil {
-				return "", execErr
-			}
-
-			t.Log("PostgresClient:", output)
-			if output == "0" {
-				return ExpectedStatus, nil
-			} else {
-				return "", fmt.Errorf("fail")
-			}
-		})
-	assert.Equal(t, ExpectedStatus, status)
-
-}
-
-func assertBambooSecrets(t *testing.T, kubectlOptions *k8s.KubectlOptions, product string) {
-	// License
-	licenseSecretName := product + "-license"
-	licenseSecretValue := getSecretValue(t, kubectlOptions, licenseSecretName, "license")
-
-	license := environment.GetFirstNonEmptyEnvVarOrFatal(t, []string{"TF_VAR_bamboo_license"})
-	assert.Equal(t, license, licenseSecretValue)
-
-	// SysAdminCredentials
-	adminSecretName := product + "-admin"
-	usernameSecretValue := getSecretValue(t, kubectlOptions, adminSecretName, "username")
-	passwordSecretValue := getSecretValue(t, kubectlOptions, adminSecretName, "password")
-	displayNameSecretValue := getSecretValue(t, kubectlOptions, adminSecretName, "displayName")
-	emailSecretValue := getSecretValue(t, kubectlOptions, adminSecretName, "emailAddress")
-
-	assert.Equal(t, "admin", usernameSecretValue)
-	assert.Equal(t, "admin", passwordSecretValue)
-	assert.Equal(t, "Admin", displayNameSecretValue)
-	assert.Equal(t, "admin@foo.com", emailSecretValue)
-
-	// Security Token
-	tokenSecretName := product + "-security-token"
-	tokenSecretValue := getSecretValue(t, kubectlOptions, tokenSecretName, "security-token")
-	assert.Equal(t, 40, len(tokenSecretValue))
-}
-
-func getSecretValue(t *testing.T, kubectlOptions *k8s.KubectlOptions, secretName string, secretKey string) (secretValue string) {
-	secret, secretErr := k8s.RunKubectlAndGetOutputE(t, kubectlOptions, "get", "secret", secretName, "-o", "jsonpath='{.data."+secretKey+"}'")
-	assert.Nil(t, secretErr)
-	assert.NotNil(t, secret)
-
-	decSecret, decodeErr := base64.StdEncoding.DecodeString(secret[1 : len(secret)-1])
-	assert.Nil(t, decodeErr)
-	secretValue = string(decSecret)
-
-	return secretValue
-}
-
-func getVpcOutput(t *testing.T, tfOptions *terraform.Options) VpcOutput {
-	vpcOutput := VpcOutput{}
-	terraform.OutputStruct(t, tfOptions, "vpc", &vpcOutput)
-	fmt.Printf("VpcOutput struct: %+v\n", vpcOutput)
-	return vpcOutput
-}
-
-func tagAsgResources(t *testing.T, environmentConfig EnvironmentConfig) {
-	AsgEc2taggingModuleTfConfig := TerraformConfig{
-		Variables: map[string]interface{}{
-			"region":           environmentConfig.TerraformConfig.Variables["region"],
-			"resource_tags":    environmentConfig.TerraformConfig.Variables["resource_tags"],
-			"environment_name": environmentConfig.TerraformConfig.Variables["environment_name"],
-			"state_type":       "local",
-			// dummy variables to pass end to end tests
-			"bamboo_license":             "test",
-			"bamboo_admin_username":      "test",
-			"bamboo_admin_password":      "test",
-			"bamboo_admin_display_name":  "test",
-			"bamboo_admin_email_address": "test",
-		},
-		EnvVariables: environmentConfig.TerraformConfig.EnvVariables,
-		TestFolder:   environmentConfig.TerraformConfig.TestFolder + "/modules/AWS/asg_ec2_tagging",
-	}
-
-	taggingModuletfOption := GenerateTerraformOptions(AsgEc2taggingModuleTfConfig, t)
-	terraform.InitAndApply(t, taggingModuletfOption)
-}
-
-func assertBambooAgentPod(t *testing.T, kubectlOptions *k8s.KubectlOptions) {
-	listPodsFilter := v1.ListOptions{
-		LabelSelector: "app.kubernetes.io/name=bamboo-agent",
-	}
-	pods := k8s.ListPods(t, kubectlOptions, listPodsFilter)
-
-	assert.Equal(t, bambooAgentCount, len(pods))
-
-	for _, pod := range pods {
-		assert.Equal(t, true, pod.Status.ContainerStatuses[0].Ready)
-	}
-}
-
 
 func bambooHealthTests(t *testing.T, testConfig TestConfig) {
 	// Test the PAUSE status
